api/pyproject.toml
[project]
name = "synkora-api"
version = "0.1.0"
description = "synkora Platform API - Production-ready LLM application platform"
authors = [{name = "synkora Team"}]
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # Web Framework
    "fastapi>=0.115.6,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "gunicorn>=23.0.0,<24.0.0",
    "python-multipart>=0.0.20,<1.0.0",
    # Database & ORM
    "sqlalchemy>=2.0.29,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "pgvector>=0.2.4,<1.0.0",
    "elasticsearch>=8.0.0,<9.0.0",
    # Data Validation & Serialization
    "pydantic[email]>=2.11.0,<3.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "orjson>=3.10.0,<4.0.0",
    # Authentication & Security
    "pyjwt>=2.10.0,<3.0.0",
    "bcrypt>=4.2.0,<5.0.0",
    "cryptography>=44.0.0,<45.0.0",
    # Task Queue
    "celery>=5.5.0,<6.0.0",
    "flower>=2.0.0,<3.0.0",
    "redis[hiredis]>=6.1.0,<7.0.0",
    # HTTP Client
    "httpx>=0.28.0,<1.0.0",
    "requests>=2.32.0,<3.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    # Utilities
    "python-dotenv>=1.0.0,<2.0.0",
    "click>=8.1.0,<9.0.0",
    "tenacity>=9.0.0,<10.0.0",
    # Google Agent SDK & MCP
    "google-genai>=1.0.0,<2.0.0",
    "fastmcp==2.13.0.2",
    "asyncpg>=0.30.0",
    # Multi-Provider LLM Support
    "litellm>=1.55.0,<2.0.0",
    "openai>=1.59.0,<2.0.0",
    "anthropic>=0.42.0,<1.0.0",
    "tiktoken>=0.9.0,<1.0.0",
    # Claude Agent SDK (for Claude Code Agent)
    "claude-agent-sdk>=0.1.44",
    # Voice Services
    "elevenlabs>=1.0.0,<2.0.0",
    "google-search-results>=2.4.2",
    "beautifulsoup4>=4.14.2",
    "pygithub>=2.8.1",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.185.0",
    "gspread>=6.0.0,<7.0.0",
    # Data Source Integrations
    "slack-sdk>=3.33.0,<4.0.0",
    "slack-bolt>=1.21.0,<2.0.0",
    "python-telegram-bot>=21.0,<22.0.0",
    # Email Services
    "sendgrid>=6.11.0,<7.0.0",
    # SSO & SAML
    # Pin xmlsec to 1.3.13 to avoid libxml2 version mismatch bug in newer versions
    "xmlsec==1.3.13",
    "python3-saml>=1.16.0,<2.0.0",
    # Web Scraping
    "crawl4ai>=0.4.0,<1.0.0",
    "playwright>=1.48.0,<2.0.0",
    "python-dateutil>=2.8.2,<3.0.0",
    # AWS S3 Storage
    "boto3>=1.35.0,<2.0.0",
    # Document Processing & Image Extraction
    "pymupdf>=1.24.0,<2.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "PyPDF2>=3.0.0,<4.0.0",
    "pillow>=10.0.0,<12.0.0",
    "python-magic>=0.4.27,<1.0.0",
    # Data Analysis
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",  # Pin to 1.x for pyarrow compatibility
    "openpyxl>=3.1.0,<4.0.0",
    "pyarrow>=14.0.0,<18.0.0",
    "databricks-sql-connector>=3.0.0,<4.0.0",
    "docker>=6.1.0,<8.0.0",
    "datadog-api-client>=2.0.0,<3.0.0",
    # Document Generation
    "reportlab>=4.0.0,<5.0.0",
    "python-pptx>=0.6.23,<1.0.0",
    "markdown>=3.5.0,<4.0.0",
    # Vector Database & Embeddings
    "qdrant-client>=1.12.0,<2.0.0",
    "pinecone-client>=5.0.0,<6.0.0",
    "sentence-transformers>=3.3.0,<4.0.0",
    "faiss-cpu>=1.8.0,<2.0.0",
    # Background Tasks
    "apscheduler>=3.10.4,<4.0.0",
    "croniter>=2.0.0,<3.0.0",
    "tzdata>=2024.1",  # IANA timezone database for zoneinfo on slim Docker images
    # Observability & Monitoring
    "prometheus-client>=0.20.0,<1.0.0",
    "langfuse>=2.58.3,<3.0.0",
    "sentry-sdk[fastapi]>=2.0.0,<3.0.0",
    "aiosqlite>=0.21.0",
    # Password Manager Integration
    "onepassword-sdk>=0.1.0",
    # YouTube Transcript Extraction
    "youtube-transcript-api>=0.6.0,<1.0.0",
    # Payment Processing
    "stripe>=11.1.0,<12.0.0",
    # WhatsApp Web / Device Link
    "neonize>=0.3.9,<1.0.0",
    "qrcode[pil]>=7.0.0,<9.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-mock>=3.14.0,<4.0.0",
    "faker>=33.0.0,<34.0.0",
    "factory-boy>=3.3.0,<4.0.0",
    
    # Code Quality
    "ruff>=0.8.0,<1.0.0",
    "mypy>=1.14.0,<2.0.0",
    "types-redis>=4.6.0",
    "types-requests>=2.32.0",
    
    # Development Tools
    "ipython>=8.31.0,<9.0.0",
    "ipdb>=0.13.0,<1.0.0",
]

llm = [
    # LLM Providers
    "openai>=1.59.0,<2.0.0",
    "anthropic>=0.42.0,<1.0.0",
    "tiktoken>=0.9.0,<1.0.0",
]

vector = [
    # Vector Databases
    "weaviate-client>=3.24.0,<4.0.0",
    "qdrant-client>=1.9.0,<2.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "asyncio: Async tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
target-version = "py311"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "DTZ", # flake8-datetimez: enforce timezone-aware datetimes (bans datetime.utcnow())
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from (too many existing violations)
    "W191",  # indentation contains tabs
    "W293",  # blank line contains whitespace (in docstrings)
    "DTZ005", # datetime.now() without tz — will fix incrementally; DTZ003 (utcnow) is the critical one
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG", "S101"]
"src/models/*.py" = ["F821"]  # Allow forward references in SQLAlchemy relationships

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = false
follow_imports = "normal"
ignore_missing_imports = true

# Disable specific error codes that are problematic with SQLAlchemy and dynamic code
disable_error_code = [
    "abstract",
    "arg-type",
    "assignment",
    "attr-defined",
    "call-arg",
    "call-overload",
    "dict-item",
    "func-returns-value",
    "import-untyped",
    "index",
    "misc",
    "name-defined",
    "no-redef",
    "operator",
    "override",
    "return-value",
    "str-bytes-safe",
    "truthy-function",
    "type-abstract",
    "type-var",
    "union-attr",
    "unused-coroutine",
    "used-before-def",
    "valid-type",
    "var-annotated",
]

[[tool.mypy.overrides]]
module = "src.models.*"
disable_error_code = ["assignment", "name-defined", "var-annotated", "return-value"]

[[tool.mypy.overrides]]
module = "src.services.*"
disable_error_code = ["arg-type", "call-overload"]

[[tool.mypy.overrides]]
module = "src.config.*"
disable_error_code = ["call-arg"]

[[tool.mypy.overrides]]
module = "src.core.*"
disable_error_code = ["dict-item", "arg-type"]
